dataflow-types cleanup, more logging
