+++ src/dataflow-types/src/types.rs
@@ -524,7 +524,7 @@ impl ExternalSourceConnector {
     }
 
     /// Returns whether or not persistence is enabled for this connector
-    pub fn enable_persistence(&self) -> bool {
+    pub fn persistence_enabled(&self) -> bool {
         match self {
             ExternalSourceConnector::Kafka(k) => k.enable_persistence,
             _ => false,
@@ -691,18 +691,6 @@ pub struct KafkaSinkConnectorBuilder {
     pub replication_factor: u32,
     pub fuel: usize,
     pub consistency_value_schema: Option<String>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-/// Data that gets sent to the persistence thread to place in persistent storage.
-/// TODO currently fairly Kafka input-centric.
-pub struct WorkerPersistenceData {
-    pub source_id: GlobalId,
-    pub partition: i32,
-    pub offset: i64,
-    pub timestamp: Timestamp,
-    pub key: Vec<u8>,
-    pub payload: Vec<u8>,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
